Extract common code to library
